updated to handle more attributes.
